+++ tommath.def
@@ -136,12 +136,11 @@ EXPORTS
     mp_sub
     mp_sub_d
     mp_submod
+    mp_to_radix
     mp_to_signed_bin
     mp_to_signed_bin_n
     mp_to_unsigned_bin
     mp_to_unsigned_bin_n
-    mp_toradix
-    mp_toradix_n
     mp_unsigned_bin_size
     mp_xor
     mp_zero
